Compile JIT sources with /W4
Commit migrated from https://github.com/dotnet/coreclr/commit/9d0f9e5443aa69525088f25d0be216261a7381a2

+++ src/coreclr/src/jit/unwindx86.cpp
@@ -21,9 +21,9 @@ XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX
 #endif // _TARGET_X86_
 
 #if defined(_TARGET_UNIX_)
-int Compiler::mapRegNumToDwarfReg(regNumber reg)
+short Compiler::mapRegNumToDwarfReg(regNumber reg)
 {
-    int dwarfReg = DWARF_REG_ILLEGAL;
+    short dwarfReg = DWARF_REG_ILLEGAL;
 
     NYI("CFI codes");
